requirements.txt
gradio[mcp]==6.0.1
httpx==0.28.1
python-dotenv==1.2.1
uvicorn==0.38.0
